Project restructure to be more in line with golang standards
- Need to improve build scripts
- Need to improve travis CI

Authored by Tyler Ramer <tramer@pivotal.io>
