+++ README.md
@@ -1,4 +1,4 @@
-# gpmt2  [![Build Status](https://travis-ci.org/pivotal-gss/gpmt2.svg?branch=master)](https://travis-ci.org/pivotal-gss/gpmt2)
+# gpmt2  [![Build Status](https://travis-ci.org/pivotal-gss/gpmt2.svg?branch=master)](https://travis-ci.org/pivotal-gss/gpmt2)  [![Go Report Card](https://goreportcard.com/badge/github.com/pivotal-gss/gpmt2)](https://goreportcard.com/report/github.com/pivotal-gss/gpmt2)
 Greenplum Magic Tool 2 - open source rebuild of GPMT
