+++ cmd/gpmt/gpmt.go
@@ -1,0 +1,35 @@
+/* Greenplum magic tool
+
+Authored by Tyler Ramer, Ignacio Elizaga
+Copyright 2018
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+
+*/
+
+package main
+
+import (
+	"os"
+
+	log "github.com/sirupsen/logrus"
+)
+
+func init() {
+	log.SetOutput(os.Stdout)
+}
+
+// the main block
+func main() {
+	Execute()
+}
