--- gpmt.go
@@ -1,31 +1,0 @@
-/* Greenplum magic tool
-
-Authored by Tyler Ramer, Ignacio Elizaga
-Copyright 2018
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-
-*/
-
-package main
-
-import (
-	"github.com/pivotal-gss/gpmt2/cmd"
-)
-
-// the main block
-func main() {
-	cmd.Execute()
-}
-
-
