--- log/logging.go
@@ -1,188 +1,0 @@
-/*
-Greenplum Magic Tool
-
-Authored by Tyler Ramer, Ignacio Elizaga
-Copyright 2018
-
-Licensed under the Apache License, Version 2.0 (the "License")
-
-*/
-package log
-
-import (
-	"github.com/Sirupsen/logrus"
-	"runtime"
-	"strings"
-	"fmt"
-	"time"
-	"os"
-	"io"
-)
-
-var logger = logrus.New()
-
-// Fields wraps logrus.Fields, which is a map[string]interface{}
-type Fields logrus.Fields
-
-// Setup the log level for logrus
-func SetLogLevel(level logrus.Level) {
-	logger.Level = level
-}
-
-// Set the logrus format
-func SetLogFormatter(formatter logrus.Formatter) {
-	logger.Formatter = formatter
-}
-
-// Enable logging to logfile.
-func SetlogFile(dest string) {
-	logFilename := fmt.Sprintf(dest + "/gpmt_log_%s", time.Now().Format("2006-01-02"))
-	logFile, err := os.OpenFile(logFilename, os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0666)
-	if err != nil {
-		Error("Could not open file: " + fmt.Sprint(err))
-		Error("Continuing with stdout only")
-	} else {
-		Info("Writing the log message to the logfile: " + logFilename)
-	}
-	mw := io.MultiWriter(os.Stdout, logFile)
-	logger.SetOutput(mw)
-}
-
-
-// Debug logs a message at level Debug on the standard logger.
-func Debug(args ...interface{}) {
-	if logger.Level >= logrus.DebugLevel {
-		entry := logger.WithFields(logrus.Fields{})
-		entry.Data["file"] = fileInfo(2)
-		entry.Debug(args...)
-	}
-}
-
-// Debug logs a message with fields at level Debug on the standard logger.
-func DebugWithFields(l interface{}, f Fields) {
-	if logger.Level >= logrus.DebugLevel {
-		entry := logger.WithFields(logrus.Fields(f))
-		entry.Data["file"] = fileInfo(2)
-		entry.Debug(l)
-	}
-}
-
-// Info logs a message at level Info on the standard logger.
-func Info(args ...interface{}) {
-	if logger.Level >= logrus.InfoLevel {
-		entry := logger.WithFields(logrus.Fields{})
-		entry.Data["file"] = fileInfo(2)
-		entry.Info(args...)
-	}
-}
-
-// Debug logs a message with fields at level Debug on the standard logger.
-func InfoWithFields(l interface{}, f Fields) {
-	if logger.Level >= logrus.InfoLevel {
-		entry := logger.WithFields(logrus.Fields(f))
-		entry.Data["file"] = fileInfo(2)
-		entry.Info(l)
-	}
-}
-
-// Warn logs a message at level Warn on the standard logger.
-func Warn(args ...interface{}) {
-	if logger.Level >= logrus.WarnLevel {
-		entry := logger.WithFields(logrus.Fields{})
-		entry.Data["file"] = fileInfo(2)
-		entry.Warn(args...)
-	}
-}
-
-// Debug logs a message with fields at level Debug on the standard logger.
-func WarnWithFields(l interface{}, f Fields) {
-	if logger.Level >= logrus.WarnLevel {
-		entry := logger.WithFields(logrus.Fields(f))
-		entry.Data["file"] = fileInfo(2)
-		entry.Warn(l)
-	}
-}
-
-// Error logs a message at level Error on the standard logger.
-func Error(args ...interface{}) {
-	if logger.Level >= logrus.ErrorLevel {
-		entry := logger.WithFields(logrus.Fields{})
-		entry.Data["file"] = fileInfo(2)
-		entry.Error(args...)
-	}
-}
-
-// Debug logs a message with fields at level Debug on the standard logger.
-func ErrorWithFields(l interface{}, f Fields) {
-	if logger.Level >= logrus.ErrorLevel {
-		entry := logger.WithFields(logrus.Fields(f))
-		entry.Data["file"] = fileInfo(2)
-		entry.Error(l)
-	}
-}
-
-// Fatal logs a message at level Fatal on the standard logger.
-func Fatal(args ...interface{}) {
-	if logger.Level >= logrus.FatalLevel {
-		entry := logger.WithFields(logrus.Fields{})
-		entry.Data["file"] = fileInfo(2)
-		entry.Fatal(args...)
-	}
-}
-
-// Debug logs a message with fields at level Debug on the standard logger.
-func FatalWithFields(l interface{}, f Fields) {
-	if logger.Level >= logrus.FatalLevel {
-		entry := logger.WithFields(logrus.Fields(f))
-		entry.Data["file"] = fileInfo(2)
-		entry.Fatal(l)
-	}
-}
-
-// Panic logs a message at level Panic on the standard logger.
-func Panic(args ...interface{}) {
-	if logger.Level >= logrus.PanicLevel {
-		entry := logger.WithFields(logrus.Fields{})
-		entry.Data["file"] = fileInfo(2)
-		entry.Panic(args...)
-	}
-}
-
-// Debug logs a message with fields at level Debug on the standard logger.
-func PanicWithFields(l interface{}, f Fields) {
-	if logger.Level >= logrus.PanicLevel {
-		entry := logger.WithFields(logrus.Fields(f))
-		entry.Data["file"] = fileInfo(2)
-		entry.Panic(l)
-	}
-}
-
-// The information of the file and the function from where the
-// the logging was thrown.
-func fileInfo(skip int) string {
-	pc, file, line, ok := runtime.Caller(skip)
-	if !ok {
-		file = "<???>"
-		line = 1
-	} else {
-		slash := strings.LastIndex(file, "/")
-		if slash >= 0 {
-			file = file[slash+1:]
-		}
-	}
-	functionName := runtime.FuncForPC(pc).Name()
-	return fmt.Sprintf("%s:%s:%d", file, functionName, line)
-}
-
-// Setup or Initialize the logger
-func init()  {
-
-	// Set the formatter option for logrus
-	formatter := new(logrus.TextFormatter)
-	formatter.TimestampFormat = "2006-01-02 15:04:05"
-	formatter.FullTimestamp = true
-
-	// Set the formatter.
-	SetLogFormatter(formatter)
-
-}
