+++ scripts/build.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+SOURCE="github.com/pivotal-gss/gpmt2/cmd/gpmt"
+TARGET="build/gpmt"
+
+echo "building binary at build/gpmt"
+
+go build -o "${TARGET}" "${SOURCE}"
